+++ src/main/java/pokeraidbot/domain/tracking/TrackingCommandListener.java
@@ -9,6 +9,8 @@ import java.util.Set;
 public interface TrackingCommandListener extends CommandListener {
     Set<PokemonTrackingTarget> getTrackingTargets(String region);
 
+    void clearCache();
+
     void add(PokemonTrackingTarget trackingTarget, User user, Config config);
 
     void remove(PokemonTrackingTarget trackingTarget, User user);